requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2==2.9.9
dash==2.14.1
plotly==5.18.0
pika==1.3.2
pytest==7.4.3
pytest-mock==3.12.0
GitPython==3.1.40
PyGithub==2.1.1
atlassian-python-api==3.41.8
jira==3.6.0
pydantic==1.10.13